+++ CHANGELOG.md
@@ -7,6 +7,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+...
+
+## [0.4.0] - 2018-03-22
+
+### Changed
+- Upgrade Clojure to 1.9.0
+- Upgrade Blocks to 1.1.0
+- Upgrade S3 SDK to 1.11.271
+
 ### Added
 - S3 Block store supports basic server-side encryption (SSE) by setting the
   `:sse` attribute to `:aes-256`.
@@ -61,7 +70,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 Initial library release.
 
-[Unreleased]: https://github.com/greglook/blocks-s3/compare/0.3.3...HEAD
+[Unreleased]: https://github.com/greglook/blocks-s3/compare/0.4.0...HEAD
+[0.4.0]: https://github.com/greglook/blocks-s3/compare/0.3.3...0.4.0
 [0.3.3]: https://github.com/greglook/blocks-s3/compare/0.3.2...0.3.3
 [0.3.2]: https://github.com/greglook/blocks-s3/compare/0.3.1...0.3.2
 [0.3.1]: https://github.com/greglook/blocks-s3/compare/0.3.0...0.3.1